requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
apscheduler>=3.10.4
pydantic-settings>=2.2.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
# test deps
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
aiosqlite>=0.20.0
